requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.3
scipy>=1.11
statsmodels>=0.14
pandas-datareader>=0.10
jupyterlab>=4.0
nbstripout>=0.6
python-dotenv>=1.0
fredapi>=0.5
